refactor: move ontology utils
